refactor :Nutrition screen
