+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/ui/screen/nutrition/NutritionScreen.kt
@@ -32,14 +32,17 @@ import com.cairosquad.evolvefit.design_system.theme.Theme
 import com.cairosquad.evolvefit.ui.component.RefreshBox
 import com.cairosquad.evolvefit.ui.navigation.NavBarRoute
 import com.cairosquad.evolvefit.ui.navigation.navBar.Scaffold
-import com.cairosquad.evolvefit.ui.screen.nutrition.component.MealHistoryItem
+import com.cairosquad.evolvefit.ui.screen.nutrition.component.MealAddedSnackBar
+import com.cairosquad.evolvefit.ui.screen.nutrition.component.MealCantAddSnackBar
+import com.cairosquad.evolvefit.ui.screen.nutrition.content.MealHistoryItem
 import com.cairosquad.evolvefit.ui.screen.nutrition.component.MealTypeDropdownMenu
 import com.cairosquad.evolvefit.ui.screen.nutrition.component.NutritionHeader
-import com.cairosquad.evolvefit.ui.screen.nutrition.component.NutritionLoadingScreen
+import com.cairosquad.evolvefit.ui.screen.nutrition.content.NutritionLoadingScreen
 import com.cairosquad.evolvefit.ui.screen.nutrition.component.NutritionSummaryCard
-import com.cairosquad.evolvefit.ui.screen.nutrition.component.SeeAll
-import com.cairosquad.evolvefit.ui.screen.nutrition.component.SuggestedMeals
-import com.cairosquad.evolvefit.ui.screen.nutrition.component.TodayMealsSummary
+import com.cairosquad.evolvefit.ui.screen.nutrition.content.NutritionContent
+import com.cairosquad.evolvefit.ui.screen.nutrition.content.SeeAll
+import com.cairosquad.evolvefit.ui.screen.nutrition.content.SuggestedMeals
+import com.cairosquad.evolvefit.ui.screen.nutrition.content.TodayMealsSummary
 import com.cairosquad.evolvefit.ui.util.ObserveAsEffect
 import com.cairosquad.evolvefit.viewmodel.nutrition.NutritionEffect
 import com.cairosquad.evolvefit.viewmodel.nutrition.NutritionInteractionListener
@@ -94,222 +97,3 @@ fun NutritionScreen(
         )
     }
 }
-
-@Composable
-private fun NutritionContent(
-    state: NutritionScreenState,
-    listener: NutritionInteractionListener,
-) {
-    RefreshBox(
-        isRefreshing = state.isRefreshing,
-        onRefresh = { listener.onRefresh() }
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .background(Theme.color.surfaces.surface)
-        ) {
-            NutritionHeader()
-            Crossfade(
-                targetState = state.screenStatus,
-                animationSpec = tween(
-                    durationMillis = 400,
-                    easing = FastOutSlowInEasing
-                )
-            ) {
-                when (state.screenStatus) {
-                    NutritionScreenState.ScreenStatus.SUCCESS -> {
-                        NutritionSuccessScreen(state, listener)
-                    }
-
-                    NutritionScreenState.ScreenStatus.LOADING -> {
-                        NutritionLoadingScreen()
-                    }
-
-                    NutritionScreenState.ScreenStatus.FAIL -> {
-                        NutritionErrorScreen { listener.onRetryClicked() }
-                    }
-                }
-            }
-        }
-        if (state.screenStatus == NutritionScreenState.ScreenStatus.SUCCESS) {
-            MealTypeDropdownMenu(
-                state = state,
-                listener = listener,
-            )
-            AddWaterIntakeBottomSheet(
-                modifier = Modifier.align(Alignment.BottomCenter),
-                listener = listener,
-                state = state
-            )
-            MealAddedSnackBar(
-                modifier = Modifier.align(Alignment.BottomCenter),
-                isVisible = state.isSnackBarVisible,
-            )
-            MealCantAddSnackBar(
-                modifier = Modifier.align(Alignment.BottomCenter),
-                isVisible = state.isSnackBarVisible,
-                state = state
-            )
-        }
-    }
-}
-
-@Composable
-private fun NutritionSuccessScreen(
-    state: NutritionScreenState,
-    listener: NutritionInteractionListener
-) {
-    LazyColumn(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(Theme.color.surfaces.surface)
-    ) {
-        item { NutritionSummaryCard(listener = listener, state = state) }
-        item { TodayMealsSummary(state = state, listener = listener) }
-        item { SuggestedMeals(state = state, listener = listener) }
-        mealHistorySection(state, listener)
-    }
-}
-
-@Composable
-private fun NutritionErrorScreen(onRetry: () -> Unit) {
-    val noInternetIcon = if (isSystemInDarkTheme()) {
-        Res.drawable.im_no_internet
-    } else {
-        Res.drawable.ic_no_internet_light
-    }
-    Box(modifier = Modifier.fillMaxSize()) {
-        StateMessage(
-            image = painterResource(noInternetIcon),
-            title = stringResource(Res.string.internet_is_not_available),
-            description = stringResource(Res.string.please_make_sure_you_are_connected_to_the_internet_and_try_again),
-            modifier = Modifier.align(Alignment.Center)
-        )
-        PrimaryButton(
-            modifier = Modifier
-                .align(Alignment.BottomCenter)
-                .padding(horizontal = 16.dp, vertical = 24.dp),
-            text = stringResource(Res.string.try_again_button),
-            enabledTextColor = Theme.color.brand.onPrimary,
-            textStyle = Theme.textStyle.body.mediumMedium14,
-            onClick = onRetry
-        )
-    }
-}
-
-private fun LazyListScope.mealHistorySection(
-    state: NutritionScreenState,
-    listener: NutritionInteractionListener
-) {
-    item {
-        SeeAll(
-            modifier = Modifier.padding(horizontal = 16.dp, vertical = 12.dp),
-            onViewAllClick = listener::onViewAllMealHistoryClicked,
-            sectionTitle = stringResource(Res.string.meal_history)
-        )
-    }
-    if (state.todayConsumedMeals.isNotEmpty()) {
-        items(state.todayConsumedMeals) { mealHistory ->
-            MealHistoryItem(meal = mealHistory)
-        }
-    } else {
-        item {
-            EmptyMealHistory()
-        }
-    }
-}
-
-@Composable
-private fun EmptyMealHistory() {
-    Box(contentAlignment = Alignment.Center) {
-        val noMealsRecorded = if (isSystemInDarkTheme()) {
-            Res.drawable.im_no_meals_recorded
-        } else {
-            Res.drawable.ic_no_meals_light
-        }
-        StateMessage(
-            modifier = Modifier.padding(vertical = 16.dp),
-            image = painterResource(noMealsRecorded),
-            title = stringResource(Res.string.no_meals_title),
-            description = stringResource(Res.string.no_meals_description)
-        )
-    }
-}
-
-@Composable
-private fun AddWaterIntakeBottomSheet(
-    state: NutritionScreenState,
-    listener: NutritionInteractionListener,
-    modifier: Modifier = Modifier
-) {
-    BottomSheet(
-        isVisible = state.isAddWaterSheetVisible,
-        onDismiss = { listener.onDismissWaterClicked() }) {
-        Column(
-            modifier = modifier
-                .padding(horizontal = 8.dp)
-                .fillMaxWidth(),
-            verticalArrangement = Arrangement.Center
-        ) {
-            Text(
-                text = stringResource(Res.string.enter_water_intake),
-                style = Theme.textStyle.title.largeBold14,
-                color = Theme.color.surfaces.onSurfaceContainer
-            )
-            Text(
-                text = stringResource(Res.string.track_water_intake),
-                style = Theme.textStyle.body.mediumMedium12,
-                color = Theme.color.surfaces.outline
-            )
-            InputField(
-                modifier = Modifier.padding(top = 16.dp),
-                value = state.consumedWaterInput,
-                onValueChange = listener::onWaterAmountChange,
-                keyboardOptions = KeyboardOptions.Default.copy(
-                    keyboardType = KeyboardType.Number
-                ),
-                placeholder = "e.g., 1.5 L",
-                leadingIcon = Res.drawable.ic_water_drop,
-                isErrorMessageShown = true,
-                error = state.waterInputError.let { it?.let { resource -> stringResource(resource) } }
-                    ?: ""
-            )
-            PrimaryButton(
-                modifier = Modifier.padding(top = 40.dp, bottom = 16.dp),
-                text = stringResource(Res.string.add_button),
-                isEnabled = state.isAddButtonEnabled,
-                onClick = {
-                    listener.onConfirmAddWaterClicked(state.consumedWaterInput)
-                })
-        }
-    }
-
-}
-
-@Composable
-private fun MealAddedSnackBar(
-    isVisible: Boolean,
-    modifier: Modifier = Modifier
-) {
-    SnackBar(
-        modifier = modifier,
-        text = stringResource(Res.string.meal_added_snackbar),
-        isVisible = isVisible
-    )
-}
-
-@Composable
-private fun MealCantAddSnackBar(
-    state: NutritionScreenState,
-    isVisible: Boolean,
-    modifier: Modifier = Modifier
-) {
-    SnackBar(
-        modifier = modifier,
-        icon = painterResource(Res.drawable.ic_info),
-        text = state.errorMessage?.let { stringResource(it) } ?: "",
-        isVisible = isVisible
-    )
-}
-
