+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/ui/screen/nutrition/content/NutritionSuccessScreen.kt
@@ -1,11 +1,15 @@
-package com.cairosquad.evolvefit.ui.screen.mealsHistory.components
+package com.cairosquad.evolvefit.ui.screen.nutrition.content
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.LazyListScope
+import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
@@ -15,21 +19,65 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.unit.dp
 import com.cairosquad.evolvefit.design_system.theme.Theme
-import com.cairosquad.evolvefit.viewmodel.meal_history.MealHistoryScreenState
+import com.cairosquad.evolvefit.ui.screen.nutrition.component.NutritionSummaryCard
+import com.cairosquad.evolvefit.viewmodel.nutrition.NutritionInteractionListener
+import com.cairosquad.evolvefit.viewmodel.nutrition.NutritionScreenState
+import com.cairosquad.evolvefit.viewmodel.nutrition.toMealIcon
 import evolvefit.composeapp.generated.resources.Res
 import evolvefit.composeapp.generated.resources.kcal_unit
+import evolvefit.composeapp.generated.resources.meal_history
 import org.jetbrains.compose.resources.painterResource
 import org.jetbrains.compose.resources.stringResource
 
+
 @Composable
-fun MealHistoryItem(
-    meal: MealHistoryScreenState.MealHistoryUiState,
+fun NutritionSuccessScreen(
+    state: NutritionScreenState,
+    listener: NutritionInteractionListener
+) {
+    LazyColumn(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(Theme.color.surfaces.surface)
+    ) {
+        item { NutritionSummaryCard(listener = listener, state = state) }
+        item { TodayMealsSummary(state = state, listener = listener) }
+        item { SuggestedMeals(state = state, listener = listener) }
+        mealHistorySection(state, listener)
+    }
+}
+
+private fun LazyListScope.mealHistorySection(
+    state: NutritionScreenState,
+    listener: NutritionInteractionListener
+) {
+    item {
+        SeeAll(
+            modifier = Modifier.padding(horizontal = 16.dp, vertical = 12.dp),
+            onViewAllClick = listener::onViewAllMealHistoryClicked,
+            sectionTitle = stringResource(Res.string.meal_history)
+        )
+    }
+    if (state.todayConsumedMeals.isNotEmpty()) {
+        items(state.todayConsumedMeals) { mealHistory ->
+            MealHistoryItem(mealHistory)
+        }
+    } else {
+        item {
+            EmptyMealHistory()
+        }
+    }
+}
+
+@Composable
+private fun MealHistoryItem(
+    meal: NutritionScreenState.ConsumedMealUiState,
     modifier: Modifier = Modifier
 ) {
     Row(
         modifier = modifier
             .fillMaxWidth()
-            .padding(vertical = 8.dp, horizontal = 16.dp),
+            .padding(vertical = 12.dp, horizontal = 16.dp),
         verticalAlignment = Alignment.CenterVertically,
         horizontalArrangement = Arrangement.SpaceBetween
     ) {
@@ -38,7 +86,7 @@ fun MealHistoryItem(
                 .clip(CircleShape)
                 .background(Theme.color.surfaces.outlineVariant)
                 .padding(10.dp),
-            painter = painterResource(meal.type.icon),
+            painter = painterResource(meal.type.toMealIcon()),
             contentDescription = null,
             tint = Theme.color.brand.primary,
         )
@@ -54,13 +102,14 @@ fun MealHistoryItem(
                 color = Theme.color.surfaces.onSurface
             )
             Text(
+                modifier = Modifier.padding(top = 8.dp),
                 text = meal.date,
                 style = Theme.textStyle.label.smallRegular12,
                 color = Theme.color.surfaces.onSurfaceVariant
             )
         }
         Column(
-            horizontalAlignment = Alignment.End
+            horizontalAlignment = Alignment.Start,
         ) {
             Text(
                 text = "${meal.calories} " + stringResource(Res.string.kcal_unit),
